# shared/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WedPlan wedding RSVP and asoebi commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/payment-config:
    get:
      operationId: getPaymentConfig
      tags: [admin]
      summary: Get safe payment configuration status
      responses:
        "200":
          description: Payment configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentConfig"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePaymentConfig
      tags: [admin]
      summary: Update payment configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentConfigInput"
      responses:
        "200":
          description: Updated payment configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentConfig"
        "400":
          description: Invalid payment configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List admin users (admin role only)
      responses:
        "200":
          description: Admin users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthUser"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAdminUser
      tags: [admin]
      summary: Create a new admin user (admin role only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminUserInput"
      responses:
        "201":
          description: Admin user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Remove an admin user (admin role only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Admin user removed
        "400":
          description: Cannot remove your own account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Admin user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    PaymentConfig:
      type: object
      required: [flutterwaveConfigured, bankTransferConfigured, flutterwaveKeyHint, webhookUrl, webhookSecretConfigured, bankTransfer]
      properties:
        flutterwaveConfigured: { type: boolean }
        flutterwaveKeyHint: { type: string }
        webhookUrl: { type: string }
        webhookSecretConfigured: { type: boolean }
        bankTransferConfigured: { type: boolean }
        bankTransfer:
          $ref: "#/components/schemas/BankTransferDetails"
    PaymentConfigInput:
      type: object
      required: [flutterwaveEnabled, flutterwaveSecretKey, bankTransferEnabled, bankTransfer]
      properties:
        flutterwaveEnabled: { type: boolean }
        flutterwaveSecretKey: { type: string }
        flutterwaveWebhookSecret: { type: string }
        bankTransferEnabled: { type: boolean }
        bankTransfer:
          $ref: "#/components/schemas/BankTransferDetails"
    BankTransferDetails:
      type: object
      required: [bankName, accountName, accountNumber, instructions]
      properties:
        bankName: { type: string }
        accountName: { type: string }
        accountNumber: { type: string }
        instructions: { type: string }
    AdminOverview:
      type: object
      required: [totalRsvps, attendingCount, asoebiInterestCount, paidOrders, recentRsvps]
      properties:
        totalRsvps: { type: number }
        attendingCount: { type: number }
        asoebiInterestCount: { type: number }
        paidOrders: { type: number }
        recentRsvps:
          type: array
          items:
            $ref: "#/components/schemas/RecentRsvp"
    RecentRsvp:
      type: object
      required: [guestName, attending, asoebiInterest, createdAt]
      properties:
        guestName: { type: string }
        attending: { type: boolean }
        asoebiInterest: { type: string, enum: [yes, no] }
        createdAt: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    AuthUser:
      type: object
      required: [id, email, role]
      properties:
        id: { type: number }
        email: { type: string }
        role: { type: string, enum: [admin, manager] }
    CreateAdminUserInput:
      type: object
      required: [email, password, role]
      properties:
        email: { type: string }
        password: { type: string, minLength: 8 }
        role: { type: string, enum: [admin, manager] }