# shared/src/generated/types/paymentConfig.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * WedPlan wedding RSVP and asoebi commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { BankTransferDetails } from './bankTransferDetails';

export interface PaymentConfig {
  flutterwaveConfigured: boolean;
  flutterwaveKeyHint: string;
  webhookUrl: string;
  webhookSecretConfigured: boolean;
  bankTransferConfigured: boolean;
  bankTransfer: BankTransferDetails;
}